# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Nami Transaction Tracking System.",
        "title": "Nami Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/vaults/{vaultId}/assets/{asset}/holdings": {
            "get": {
                "description": "Get derived holdings for a specific asset/account within a vault",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "Get asset holdings in a vault (derived)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset symbol",
                        "name": "asset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "account",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/holdings": {
            "get": {
                "description": "Get derived vault holdings from transaction ledger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "Get vault holdings (derived)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/transactions": {
            "get": {
                "description": "List transactions for a vault (supports limit, offset, type, status)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "List vault transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VaultTransaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new transaction recorded in the vault ledger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "Create a vault transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vault transaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VaultTransaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VaultTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/user/{userId}/holdings": {
            "get": {
                "description": "Get derived user holdings in a vault from transaction ledger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "Get user holdings (derived)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/user/{userId}/transactions": {
            "get": {
                "description": "List transactions for a user in a vault (supports limit, offset)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "List user transactions in a vault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VaultTransaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.VaultTransaction": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "amount_usd": {
                    "description": "Financial amounts",
                    "type": "number"
                },
                "asset": {
                    "description": "Asset information (for rebalancing and asset-specific transactions)",
                    "type": "string"
                },
                "asset_price": {
                    "type": "number"
                },
                "asset_quantity": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "Audit fields",
                    "type": "string"
                },
                "executed_at": {
                    "type": "string"
                },
                "external_tx_id": {
                    "type": "string"
                },
                "fee_amount": {
                    "description": "Fee information",
                    "type": "number"
                },
                "fee_rate": {
                    "type": "number"
                },
                "fee_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "price_per_share": {
                    "type": "number"
                },
                "share_price_after": {
                    "type": "number"
                },
                "share_price_before": {
                    "type": "number"
                },
                "shares": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Metadata",
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "type": {
                    "description": "Transaction details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.VaultTransactionType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_shares_after": {
                    "type": "number"
                },
                "user_shares_before": {
                    "description": "User balance snapshots",
                    "type": "number"
                },
                "vault_aum_after": {
                    "type": "number"
                },
                "vault_aum_before": {
                    "description": "Price and valuation information",
                    "type": "number"
                },
                "vault_id": {
                    "type": "string"
                }
            }
        },
        "models.VaultTransactionType": {
            "type": "string",
            "enum": [
                "deposit",
                "withdrawal",
                "mint_shares",
                "burn_shares",
                "rebalance",
                "fee",
                "yield",
                "income",
                "expense",
                "valuation"
            ],
            "x-enum-varnames": [
                "VaultTxTypeDeposit",
                "VaultTxTypeWithdrawal",
                "VaultTxTypeMintShares",
                "VaultTxTypeBurnShares",
                "VaultTxTypeRebalance",
                "VaultTxTypeFee",
                "VaultTxTypeYield",
                "VaultTxTypeIncome",
                "VaultTxTypeExpense",
                "VaultTxTypeValuation"
            ]
        }
    }
}